requirements.txt
nltk==3.8.1
python-telegram-bot==13.7
python-dotenv==0.21.1
